app/requirements.txt
fastapi==0.115.10
uvicorn==0.34.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

